[KYH] 예약 승인 및 거절
[KYH] 예약 승인 및 거절
