+++ plend/src/main/java/three/people/service/HostServiceImpl.java
@@ -214,6 +214,18 @@ public class HostServiceImpl implements HostService {
 	public int approval(BookVO bookVO) {
 		return hostDAO.approval(bookVO);
 	}
+
+
+	@Override
+	public int insertReject(BookVO bookVO) {
+		int result = hostDAO.insertReject(bookVO);
+			if(result == 1) {
+				System.out.println("거절 사유 입력 완료");
+			}else {
+				System.out.println("거절 사유 입력 실패 ");
+			}
+		return result;
+	}
