+++ plend/src/main/webapp/WEB-INF/views/host/ajax/bookList.jsp
@@ -2,6 +2,7 @@
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
 <%@ page session="true" %>
 
+
 <table class="table table-hover text-center clearfix" >
 	<thead class="table-dark">
 		<tr style="text-al">
@@ -31,16 +32,15 @@
 			<td>${pv.useTime}</td>
 			<td>${pv.cntPeople}</td>
 			<td>${pv.nickName}</td>
-			<c:choose>
+			 <c:choose>
 				<c:when test="${pv.approvalYN eq 'N'}">
-					<td>승인 대기</td>
+					<td id="approvalY${pv.bidx}">승인 대기</td>
 				</c:when>
-				<c:when test="${pv.approvalYN eq 'Y'}">
-					<td id="approvalY">승인 완료</td>
+				<c:when test="${pv.approvalYN eq 'Y'}"> 
+					<td >승인 완료</td>
 				</c:when>
-			</c:choose>
-			
-			<td id="canclebtn${pv.bidx}"><button class="btnDefault" type="button" onclick="approval(${pv.bidx},'${pv.approvalYN}')" id="approvalBtn${pv.bidx}">확인</button></td>
+			</c:choose> 
+			<td id="canclebtn${pv.bidx}"><button class="btnDefault" type="button" onclick="approval(${pv.bidx})" id="approvalBtn${pv.bidx}">확인</button></td>
 			<!-- 2022.07.15 김연희:
 			 확인 버튼 누르면 예약정보 상세창이 팝업으로 뜨고  승인, 거절 버튼이 주어지고 
 			 승인을 누르면 approval값 변경. 
@@ -111,39 +111,7 @@
 	}
 	</script>
 	
-	<script>
-		function approval(idx, app){
-			
-			if(app == 'Y'){
-				alert("이미 승인된 요청입니다.")
-			}else{
-				
-				if(confirm("승인하시겠습니까? 승인 완료 후 예약을 취소 할 수 없습니다.")){
-				$.ajax({
-					url: "approval.do", 
-					type: "GET",
-					data: "bidx="+idx,
-					success: function(){
-						alert("승인되었습니다.")
-						$("#approvalY").html("승인 완료");
-					}
-				});
-				
-				}else{
-				console.log("승인 거절");
-				}
-			
-			}
-			
-				
-			
-			
-		}
-	
-	
-	
-	
-	</script>
+
